redistributed logic among files
